Fix handling of multiple blocks
It appears that
if we modify the AST with ReplaceChild
while walking it with ast.Walk,
we may end up skipping over sibling nodes.

For example per the report in #3,
with two mermaid blocks next to each other,
we end up replacing the first and skipping the second.

Fix this by collecting all mermaid nodes in one pass,
and replacing them in the second.

Resolves #3