Support for server-side rendering
Adds a new ServerRenderer which generates SVG from Mermaid blocks
by shelling out to the 'mmdc' executable (if found)
and replacing the code blocks with the SVG inline.

The following is the surface for the new renderer:

```go
type ServerRenderer struct {
  MMDC MMDC
  // ...
}

type MMDC interface{ /* ... */ }
  var DefaultMMDC // unexported type
  type CLI struct // implements MMDC
```

The Extender picks this automatically if 'mmdc' is found on PATH.
This is customized with RenderMode:

```diff
 type Extender struct {
+  RenderMode RenderMode
   // ...
 }
```

Valid values for RenderMode are: Auto (default), Client, Server.

---

Testing:
Besides adding an integration test,
this includes a new internal/exectest package
that lets tests supply behavior for external binaries.
This is a usability improvement over TestMain hijacking.
It lets the test and the test behavior reside next to each other.