Add CDP-based server-side rendering
Introduces support for rendering Mermaid diagrams server-side
with the help of a long-running headless Chromium-based browser.
This talks to the browser using Chrome DevTools Protocol,
invoking MermaidJS, and getting SVGs back from it.
This is the same method used by the MermaidJS CLI.
This method opens up room in the future for non-SVG diagrams as well.

To make this possible,
the server-side renderer was decoupled from CLI-based rendering.
It now exposes a Compiler option which can be CLI or CDP based.

To reduce some noise in the namespace,
the `MMDC` type was renamed to `CLI`,
and the old `CLI` type was hidden away.

Resolves #28