test(integration): Fix not being run, use latest Mermaid
The integration test for server-side prudently added a skip if 'mmdc'
wasn't on $PATH.
However, with the conditons it runs under, mmdc is never on PATH.
It's under node_modules/.bin.
So this was just always being skipped.

Re-enable the test, and upgrade it to the latest version of MermaidJS.